LNbits safety notice
The LNbits Super Admin Wallet contains very sensitive data. It is not recommended to open or even save this in a normal browser. It is better to save and open it separately in an incognito browser. The wallet data can be restored at any time by the owner of the LNbits wallet.